+++ App/DAL/Entities/ScrumTask.cs
@@ -8,8 +8,8 @@ namespace App.DAL.Entities
     public class ScrumTask:Task
     {
         public DateTime ExpirationDate { get; set; }
-        public double StoryPointsExpected { get; set; }
-        public double StoryPointsSpent { get; set; }
+        public int StoryPointsExpected { get; set; }
+        public int StoryPointsSpent { get; set; }
 
         public virtual Sprint Sprint { get; set; }
         public int SprintId { get; set; }
